requirements/base.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-js-asset==2.0.0
django-mptt==0.13.4
django-mptt-admin==2.2.0
docutils==0.18.1
idna==3.3
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
